NUMICRO: Moved the serial code to UARTv1
The next NUMICRO MCU that will run ChibiOS changes many register names.
Some of them look like they are used identical to the old registers,
but just renamed. On some occasions it would be feasible to only redefine
the names, but the code won't be more comprehensible when read side by side
to the technical reference manual. Moreover there will be some minor
changes to the usage as well, so going for versioned drivers like the
STM32 port seems like a good idea now.